models and utils reorganized, backup typings
