+++ src/utils/index.ts
@@ -5,8 +5,9 @@ import Waitress from './waitress';
 import EqualsPartial from './equalsPartial';
 import RealpathSync from './realpathSync';
 import AssertString from './assertString';
+import * as BackupUtils from './backup';
 
 export {
     Wait, IsNumberArray, Queue, Waitress, EqualsPartial, RealpathSync,
-    AssertString,
+    AssertString, BackupUtils, 
 };
